Added public methods to memory to get concept priority, desire value, truth value. Refactored package structure
