+++ nars_core_java/nars/mental_operator/Anticipate.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package nars.Operator_Mental;
+package nars.mental_operator;
