+++ nars_core_java/nars/mental_operator/Fire.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package nars.Operator_Mental;
+package nars.mental_operator;
 
 import java.util.ArrayList;
 import nars.entity.Task;
@@ -16,10 +16,10 @@ import nars.storage.Memory;
  *
  * @author liyiy
  */
-public class Escape extends Operator{
-
-    public Escape() {
-        super("^escape");        
+public class Fire extends Operator{
+    
+    public Fire() {
+        super("^fire");        
     }
 
     public void escapeFeedback(final Term content, final Task t, final Memory memory, final long time) {
